[refactor]: - types - queries - mutations
