+++ src/components/persons/Persons.tsx
@@ -1,20 +1,21 @@
-import { useSearchPerson } from "../../queries/searchPerson";
+import { useSearchPerson } from "../../hooks/useSearchPerson";
 import { CreatePersonForm } from "./form/CreatePerson";
 import { PersonDetail } from "./PersonDetail";
 
-//$name: String!, $street:String!, $city:String!, $phone:String
-
-interface AllPersons {
-  name: string;
-  phone: string;
-  address?: {
-    city?: string;
-    street?: string;
-  };
+interface Address {
+  city: string;
+  street: string;
 }
+
+export interface Person {
+  name: string;
+  phone?: string;
+  address: Address;
+}
+
 interface PersonProps {
   data: {
-    allPersons: AllPersons[];
+    allPersons: Person[];
   };
 }
