+++ src/graphql/mutations.ts
@@ -1,0 +1,20 @@
+import { gql } from "@apollo/client";
+
+export const CREATE_PERSON = gql`
+  mutation createPerson(
+    $name: String!
+    $street: String!
+    $city: String!
+    $phone: String
+  ) {
+    addPerson(name: $name, phone: $phone, street: $street, city: $city) {
+      name
+      phone
+      address {
+        city
+        street
+      }
+      id
+    }
+  }
+`;
