+++ src/graphql/queries.ts
@@ -1,0 +1,28 @@
+import { gql } from "@apollo/client";
+
+export const ALL_PERSONS = gql`
+  query {
+    allPersons {
+      name
+      phone
+      address {
+        city
+        street
+      }
+    }
+  }
+`;
+
+export const FIND_PERSON = gql`
+  query findPersonByName($nameToSearch: String!) {
+    findPerson(name: $nameToSearch) {
+      name
+      phone
+      id
+      address {
+        street
+        city
+      }
+    }
+  }
+`;
