+++ src/hooks/useSearchPerson.ts
@@ -1,18 +1,7 @@
 import { useLazyQuery } from "@apollo/client";
 import { useEffect, useState } from "react";
 import { FIND_PERSON } from "../graphql/queries";
-
-// TODO: refactor the types
-interface Address {
-  street?: string;
-  city?: string;
-}
-export interface Person {
-  name: string;
-  phone: string;
-  id: string;
-  address: Address;
-}
+import { Person } from "../components/persons/Persons";
 
 export const useSearchPerson = () => {
   const [getPerson, result] = useLazyQuery(FIND_PERSON);
